// src/api/or-planner/models/OperatingRoom.ts
export interface OperatingRoom {
    /**
     * 
     * @type {string}
     * @memberof OperatingRoom
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof OperatingRoom
     */
    number: string;
    /**
     * 
     * @type {string}
     * @memberof OperatingRoom
     */
    name: string;
    /**
     * general, cardiac, orthopedic, etc.
     * @type {string}
     * @memberof OperatingRoom
     */
    type?: string;
    /**
     * 
     * @type {number}
     * @memberof OperatingRoom
     */
    capacity?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof OperatingRoom
     */
    equipment?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof OperatingRoom
     */
    status: OperatingRoomStatusEnum;
    /**
     * 
     * @type {Array<ScheduledOperation>}
     * @memberof OperatingRoom
     */
    scheduledOperations?: Array<ScheduledOperation>;
    /**
     * 
     * @type {Array<ProcedureType>}
     * @memberof OperatingRoom
     */
    predefinedProcedures?: Array<ProcedureType>;
    /**
     * 
     * @type {boolean}
     * @memberof OperatingRoom
     */
    archived?: boolean;
}


/**
 * @export
 */
export const OperatingRoomStatusEnum = {
    Active: 'active',
    Maintenance: 'maintenance',
    Inactive: 'inactive'
} as const;
export type OperatingRoomStatusEnum = typeof OperatingRoomStatusEnum[keyof typeof OperatingRoomStatusEnum];