// src/api/or-planner/models/Patient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OR Planner API
 * Operating Room Planner API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: filip.chromek@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Patient
 */
export interface Patient {
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    lastName: string;
    /**
     * Slovak birth number, slash optional
     * @type {string}
     * @memberof Patient
     */
    birthNumber: string;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    contact?: string;
    /**
     * VšZP, Dôvera, Union
     * @type {string}
     * @memberof Patient
     */
    insurance?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Patient
     */
    archived?: boolean;
}

/**
 * Check if a given object implements the Patient interface.
 */
export function instanceOfPatient(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "firstName" in value;
    isInstance = isInstance && "lastName" in value;
    isInstance = isInstance && "birthNumber" in value;

    return isInstance;
}

export function PatientFromJSON(json: any): Patient {
    return PatientFromJSONTyped(json, false);
}

export function PatientFromJSONTyped(json: any, ignoreDiscriminator: boolean): Patient {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'birthNumber': json['birthNumber'],
        'contact': !exists(json, 'contact') ? undefined : json['contact'],
        'insurance': !exists(json, 'insurance') ? undefined : json['insurance'],
        'archived': !exists(json, 'archived') ? undefined : json['archived'],
    };
}

export function PatientToJSON(value?: Patient | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'firstName': value.firstName,
        'lastName': value.lastName,
        'birthNumber': value.birthNumber,
        'contact': value.contact,
        'insurance': value.insurance,
        'archived': value.archived,
    };
}